+++ src/library/inputs/xinput.cpp
@@ -72,8 +72,145 @@ int XISelectEvents(Display* dpy, Window win, XIEventMask *masks, int num_masks)
 XIDeviceInfo* XIQueryDevice(Display* dpy, int deviceid, int* ndevices_return)
 {
     DEBUGLOGCALL(LCF_WINDOW);
-    LINK_NAMESPACE_FULLNAME(XIQueryDevice, "libXi.so.6");
-    return orig::XIQueryDevice(dpy, deviceid, ndevices_return);
+    
+    int device_count = 0;
+    switch (deviceid) {
+        case XIAllDevices:
+        case XIAllMasterDevices:
+            device_count = 2;
+            break;
+        case 2:
+        case 3:
+            device_count = 1;
+            break;
+        default:
+            device_count = 0;
+            break;
+    }
+
+    *ndevices_return = device_count;
+
+    if (device_count == 0) {
+        /* TODO: Generate BadDevice error */
+        return nullptr;
+    }
+
+    XIDeviceInfo* info = new XIDeviceInfo[device_count+1];
+    int d = 0;
+
+    if (deviceid != 3) {
+        /* Build pointer device info. Values come mostly from my mouse, while
+        * keeping only the necessary number of inputs. */
+        info[d].deviceid = 2;
+        info[d].name = "Virtual core pointer";
+        info[d].use = XIMasterPointer;
+        info[d].attachment = 3;
+        info[d].enabled = True;
+        info[d].num_classes = 3;
+        info[d].classes = new XIAnyClassInfo*[info[d].num_classes];
+        
+        XIButtonClassInfo* bc = new XIButtonClassInfo;
+        info[d].classes[0] = reinterpret_cast<XIAnyClassInfo*>(bc);
+        bc->type = XIButtonClass;
+        bc->sourceid = info[d].deviceid;
+        bc->num_buttons = 5;
+        bc->labels = new Atom[bc->num_buttons];
+        NATIVECALL(bc->labels[0] = XInternAtom(dpy, "Button Left", False));
+        NATIVECALL(bc->labels[1] = XInternAtom(dpy, "Button Middle", False));
+        NATIVECALL(bc->labels[2] = XInternAtom(dpy, "Button Right", False));
+        NATIVECALL(bc->labels[3] = XInternAtom(dpy, "Button Wheel Up", False));
+        NATIVECALL(bc->labels[4] = XInternAtom(dpy, "Button Wheel Down", False));
+        bc->state.mask_len = 0;
+        bc->state.mask = nullptr;
+        
+        XIValuatorClassInfo* v1 = new XIValuatorClassInfo;
+        info[d].classes[1] = reinterpret_cast<XIAnyClassInfo*>(v1);
+        v1->type = XIValuatorClass;
+        v1->sourceid = info[d].deviceid;
+        v1->number = 0;
+        NATIVECALL(v1->label = XInternAtom(dpy, "Rel X", False));
+        v1->min = -1;
+        v1->max = -1;
+        v1->value = -1;
+        v1->resolution = 0;
+        v1->mode = XIModeRelative;
+        
+        XIValuatorClassInfo* v2 = new XIValuatorClassInfo;
+        info[d].classes[2] = reinterpret_cast<XIAnyClassInfo*>(v2);
+        v2->type = XIValuatorClass;
+        v2->sourceid = info[d].deviceid;
+        v2->number = 0;
+        NATIVECALL(v2->label = XInternAtom(dpy, "Rel Y", False));
+        v2->min = -1;
+        v2->max = -1;
+        v2->value = -1;
+        v2->resolution = 0;
+        v2->mode = XIModeRelative;
+        
+        d++;
+    }
+
+    if (deviceid != 3) {
+        /* Build keyboard device info. */
+        info[d].deviceid = 3;
+        info[d].name = "Virtual core keyboard";
+        info[d].use = XIMasterKeyboard;
+        info[d].attachment = 2;
+        info[d].enabled = True;
+        info[d].num_classes = 1;
+        info[d].classes = new XIAnyClassInfo*[info[d].num_classes];
+        
+        XIKeyClassInfo* kc = new XIKeyClassInfo;
+        info[d].classes[0] = reinterpret_cast<XIAnyClassInfo*>(kc);
+        kc->type = XIKeyClass;
+        kc->sourceid = info[d].deviceid;
+        kc->num_keycodes = 248;
+        kc->keycodes = new int[kc->num_keycodes];
+        for (int i=0; i<kc->num_keycodes; i++) {
+            kc->keycodes[i] = i+8;
+        }
+        
+        d++;
+    }
+    
+    /* Last dummy item to indicate the end of the list for XIFreeDeviceInfo() */
+    info[d].deviceid = 0;
+
+    return info;
+}
+
+void XIFreeDeviceInfo( XIDeviceInfo *info)
+{
+    DEBUGLOGCALL(LCF_WINDOW);
+    
+    if (!info)
+        return;
+        
+    for (int d = 0; info[d].deviceid != 0; d++) {
+        for (int c = 0; c < info[d].num_classes; c++) {
+            switch (info[d].classes[c]->type) {
+                case XIButtonClass: {
+                    XIButtonClassInfo* bc = reinterpret_cast<XIButtonClassInfo*>(info[d].classes[c]);
+                    delete bc->labels;
+                    delete bc;
+                    break;
+                }
+                case XIValuatorClass: {
+                    XIValuatorClassInfo* vc = reinterpret_cast<XIValuatorClassInfo*>(info[d].classes[c]);
+                    delete vc;
+                    break;
+                }
+                case XIKeyClass: {
+                    XIKeyClassInfo* kc = reinterpret_cast<XIKeyClassInfo*>(info[d].classes[c]);
+                    delete kc->keycodes;
+                    delete kc;
+                    break;
+                }                
+            }
+        }
+        delete info[d].classes;
+    }
+    delete info;    
 }
 
 }