+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add new SDL audio functions and update SDL_dynapi (#518)
 * RAM search can be interrupted
 * Hook mono sleep function
+* Implement XIQueryDevice()
 
 ### Changed